feat: migrate test suite from jest/ts-jest to vitest
- Replace jest, ts-jest, @types/jest with vitest and @vitest/coverage-v8
- Create vitest.config.ts from jest.config.cjs settings
  - globals: true for implicit test globals (describe, it, expect)
  - resolve.alias for .js extension stripping (moduleNameMapper equivalent)
  - setupFiles pointing to vitest.setup.ts
- Rename src/jest.setup.ts to src/vitest.setup.ts
  - Update custom matcher type augmentation from jest namespace to vitest module
  - toMatchStructure custom matcher preserved as-is (compatible API)
- Update src/helpers/__tests__/repeat.spec.ts: jest.fn() -> vi.fn()
- Update package.json scripts: remove NODE_OPTIONS, jest -> vitest
- Remove all jest-related resolutions (no longer needed)
- Add vitest/globals to tsconfig.json types for global type support
- Delete jest.config.cjs
- All 279 tests pass (277 passed, 2 skipped - same as jest)
- Coverage verified with v8 provider
