+++ package.json
@@ -7,9 +7,9 @@
     "module": "./dist/esm/index.js",
     "types": "./types/index.d.ts",
     "scripts": {
-        "test": "NODE_OPTIONS='--experimental-vm-modules' jest",
-        "test:inspect": "NODE_OPTIONS='--experimental-vm-modules --inspect' node node_modules/.bin/jest",
-        "test:coverage": "NODE_OPTIONS='--experimental-vm-modules' jest --coverage",
+	"test": "vitest run",
+	"test:inspect": "vitest run --inspect",
+	"test:coverage": "vitest run --coverage",
         "clear": "run-script-os",
         "clear:win32": "rmdir /s /q dist && rmdir /s /q types",
         "clear:default": "(rm -rf ./dist || exit 0) && (rm -rf ./types || exit 0)",
@@ -54,35 +54,25 @@
         "typescript": "6.0.3"
     },
     "devDependencies": {
-        "@types/jest": "^30.0.0",
-        "jest": "^30.4.2",
+		"@vitest/coverage-v8": "^3.2.1",
+		"vitest": "^3.2.1",
         "madge": "^8.0.0",
         "run-script-os": "^1.1.6",
-        "ts-jest": "^29.4.9",
-        "ts-node": "^10.9.2",
+	"ts-node": "^10.9.2",
         "tsx": "^4.21.0",
         "typedoc": "^0.28.19",
         "uuid": "^14.0.0",
         "webpack": "^5.106.2",
         "webpack-cli": "^7.0.2"
     },
-    "resolutions": {
-        "jest/**/anymatch/picomatch": "^2.3.2",
-        "@types/jest/**/micromatch/picomatch": "^2.3.2",
-        "**/jest-haste-map/picomatch": "^4.0.4",
-        "**/jest-message-util/picomatch": "^4.0.4",
-        "**/jest-util/picomatch": "^4.0.4",
-        "glob/minimatch": "^9.0.7",
-        "typedoc/minimatch": "^9.0.7",
-        "madge/**/@typescript-eslint/typescript-estree/minimatch": "^9.0.7",
-        "jest/**/test-exclude/minimatch": "^3.1.3",
-        "jest/**/test-exclude/glob/minimatch": "^3.1.3",
-        "madge/**/glob/minimatch": "^3.1.3",
-        "minimatch/brace-expansion": "^2.1.0",
-        "jest/**/test-exclude/minimatch/brace-expansion": "^1.1.14",
-        "madge/**/glob/minimatch/brace-expansion": "^1.1.14",
-        "webpack/**/ajv/fast-uri": "^3.1.2"
-    },
+	"resolutions": {
+		"glob/minimatch": "^9.0.7",
+		"typedoc/minimatch": "^9.0.7",
+		"madge/**/@typescript-eslint/typescript-estree/minimatch": "^9.0.7",
+		"madge/**/glob/minimatch": "^3.1.3",
+		"madge/**/glob/minimatch/brace-expansion": "^1.1.14",
+		"webpack/**/ajv/fast-uri": "^3.1.2"
+	},
     "exports": {
         ".": {
             "types": "./types/index.d.ts",
